requirements.txt
pylint
plotly.express
dash
dash_core_components
dash_html_components
pandas
sqlalchemy